Add new `LBAL_CONFIG_cpp_version` token
Rather than add special-case code everywhere to test for whether
we’re running under MSVS and detect the C++ dialect appropriately,
there’s now a global token set to mirror what `__cplusplus` _should_
be, if compilers were all conformant. Note that we still punt in the case
of compilers self-identifying as using a dialect newer than the known
official dialects, e.g., following the convention of incrementing the
latest official value by 1 to indicate partial or full conformance with an
draft Standard; in this case, we just track the supplied value rather than
fall back to a well-known one.