Front-end refactor to reduce public header count
- eliminated `lbalVersion.hpp`
- moved `lbalFeatureSetup.hpp` to `details/`
- updated references in comments and docs
- removed header file references and source from docs

Note that `lbalConfig.hpp` remains as a public header,
for now. It will eventually also be moved to `details/`
once an override mechanism is worked out (e.g.,
assuming definition of override values on the
command line as arguments to the build system, or
via inclusion of a user-suppliedlocal header if one is
found, etc.).